Smol refractures
